quickstart-v2/requirements.txt
streamlit
pandas
altair
google-cloud-firestore
google-auth
python-dotenv
protobuf
google-cloud-bigquery
altair
protobuf
vl-convert-python
streamlit_cookies_manager
google-cloud-logging
google-cloud-secret-manager
google_auth_oauthlib
google-cloud-geminidataanalytics